Moving validate function and type into place.
