+++ graylog2-web-interface/src/views/components/searchbar/queryvalidation/types.ts
@@ -1,0 +1,11 @@
+export type QueryValidationState = {
+  status: 'OK' | 'ERROR' | 'WARNING',
+  explanations: Array<{
+    errorType: string,
+    errorMessage: string,
+    beginLine: number,
+    endLine: number,
+    beginColumn: number,
+    endColumn: number,
+  }> | undefined
+}
